fix: поправил верстку + обновил пути
