--- src/pages/MainPage/index.ts
@@ -1,3 +1,0 @@
-export {
-  MainPageAsync as MainPage,
-} from './ui/MainPage.async';
