--- src/pages/MainPage/ui/MainPage.async.tsx
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const MainPageAsync = lazy(() => import('./MainPage'));
