--- src/pages/MainPage/ui/MainPage.tsx
@@ -1,9 +1,0 @@
-const MainPage = () => {
-  return (
-    <div>
-      Главная страница
-    </div>
-  );
-};
-
-export default MainPage;
