+++ src/pages/Onboarding/index.ts
@@ -1,0 +1,1 @@
+export { Onboarding } from './ui/Onboarding';
