+++ src/pages/Onboarding/ui/Onboarding.module.css
@@ -1,22 +1,27 @@
-/* 
-    1. нужно будет пофиксить gradient, + добавить keyframes для slide.
-*/
-
-@import url('../../../app/styles/variables/global.css');
-
 .onboard {
-    max-width: 390px;
+    /* max-width: 390px;
     margin: 0 auto;
-    padding: 0;
+    padding: 0; */
+    width: 100%;
+    height: 100vh;
 }
 
 .container {
     width: 100%;
     margin: 0 auto;
     position: relative;
+    display: flex;
+    flex-direction: column-reverse;
+    align-items: center;
+    gap: 75px;
     background-position: top;
-    background-size: 100% 100%;
-    height: 844px;
+    background-size: cover;
+    background-repeat: no-repeat;
+    height: 100%;
+}
+
+.fourthSlide {
+    background-position: right 50% calc(100% - 250px) -250px;
 }
 
 .gradient {
@@ -25,20 +30,19 @@
     left: 0;
     right: 0;
     height: calc(844px - 372px); 
-    background: linear-gradient(0deg, rgba(21, 21, 21, 0.90) 50.95%, rgba(21, 21, 21, 0.00) 99.7%);
+    background: linear-gradient(0deg, rgb(21 21 21 / 90%) 50.95%, rgb(21 21 21 / 0%) 99.7%);
     background-blend-mode: multiply; /* здесь надо будет св-во поменять, так как в figma св-ва нет */
     z-index: 1; 
 }
 
 .info {
-    margin: 0 auto;
-    position: relative;
-    padding-top: 519px;
     display: flex;
     flex-direction: column;
     text-align: center;
-    z-index: 3;
     max-width: 350px;
+    gap: 16px;
+    flex: 0 1 126px;
+    line-height: 1.25;
 }
 
 .title {
@@ -52,40 +56,21 @@
 }
 
 .subtitle {
-    margin: 16px 0 0 0;
+    margin: 0;
     padding: 0;
     font-family: var(--font-rubik-semibold);
     color: var(--secondary-font-color);
     font-size: 16px;
     max-width: 350px;
-    text-align: center;
     width: 100%;
     letter-spacing: -0.02em;
-}
-
-.close {
-    width: 24px;
-    height: 24px;
-    position: absolute;
-    top: 82px;
-    right: 17px;
-    background-size: 100% 100%;
-    background-repeat: no-repeat;
-    border: none;
-    background-color: transparent;
-    z-index: 3;
-    cursor: pointer;
 }
 
 .slider {
     max-width: 248px;
     display: flex;
     gap: 12px;
-    position: relative;
-    top: 68px;
-    z-index: 3;
-    left: 50%;
-    transform: translateX(-50%);
+    padding-bottom: 9.4vh;
 }
 
 .slide {
