--- src/pages/TestPage/index.ts
@@ -1,3 +1,0 @@
-export {
-  TestPageAsync as TestPage,
-} from './ui/TestPage.async';
