--- src/pages/TestPage/ui/TestPage.async.tsx
@@ -1,3 +1,0 @@
-import { lazy } from 'react';
-
-export const TestPageAsync = lazy(() => import('./TestPage'));
