--- src/pages/TestPage/ui/TestPage.tsx
@@ -1,9 +1,0 @@
-const TestPage = () => {
-  return (
-    <div>
-      Тестовая страница
-    </div>
-  );
-};
-
-export default TestPage;
