+++ src/shared/config/routeConfig/routeConfig.tsx
@@ -1,31 +1,17 @@
 import { RouteProps } from 'react-router-dom';
-import { MainPage } from 'pages/MainPage';
-import { TestPage } from 'pages/TestPage';
-import { Onboarding } from '../../../widgets/Onboarding/index'
+import { Onboarding } from '../../../pages/Onboarding/index';
 
 export enum AppRoutes {
-  MAIN = 'main',
-  TEST = 'test',
-  ONBOARDING = 'test2'
+  ONBOARDING = 'main',
 }
 
 export const RoutePath: Record<AppRoutes, string> = {
-  [AppRoutes.MAIN]: '/',
-  [AppRoutes.TEST]: '/test',
-  [AppRoutes.ONBOARDING]: '/test2',
+  [AppRoutes.ONBOARDING]: '/',
 };
 
 export const routeConfig: Record<AppRoutes, RouteProps> = {
-  [AppRoutes.MAIN]: {
-    path: RoutePath.main,
-    element: <MainPage />,
-  },
-  [AppRoutes.TEST]: {
-    path: RoutePath.test,
-    element: <TestPage />,
-  },
   [AppRoutes.ONBOARDING]: {
-    path: RoutePath.test2,
-    element: <Onboarding /> 
-  }
+    path: RoutePath.main,
+    element: <Onboarding />,
+  },
 };
