--- src/widgets/Onboarding/index.ts
@@ -1,1 +1,0 @@
-export { Onboarding } from './ui/Onboarding';
